enable to show the completion tooltip.
