+++ bin/release/setting/editor-setting.xml
@@ -11,7 +11,7 @@
         <b>255</b>
     </textColor>
     <textFont>
-        <family>00000242601DE7F0</family>
+        <family>000001EFF28301A0</family>
         <pointSize>14</pointSize>
         <weight>400</weight>
         <italic>false</italic>
